requirements.txt
acoustics==0.2.6
aiohttp==3.13.2
audiomentations==0.43.1
datasets==4.4.1
matplotlib==3.10.8
mutagen==1.47.0
numpy==2.1.2
onnxruntime==1.23.2
pronouncing==0.2.0
pytest==9.0.2
pytest-cov==7.0.0
pytest-flake8==1.3.0
pytest-mypy==1.0.1
requests==2.32.5
resampy==0.4.3
scikit-learn==1.7.2
scipy==1.15.3
soundfile==0.13.1
sounddevice==0.5.3
speechbrain==1.0.3
speexdsp_ns==0.1.2
torchcodec==0.2.1
torch-audiomentations==0.12.0
tqdm==4.67.1
